requirements-full.txt
# =======================================================
# AI HEALTHCARE - FULL REQUIREMENTS (Local Development)
# =======================================================
# Use for local dev with full ML capabilities
# For Render deployment, use requirements.txt (lite version)
# =======================================================

# --- Core API & Server ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
orjson
starlette
starlette
requests>=2.32.3
razorpay

# --- Authentication & Security ---
passlib[bcrypt]>=1.7.4
bcrypt<4.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0

# --- Database & Config ---
sqlalchemy
psycopg2-binary
python-dotenv

# --- Data Science & ML (Full) ---
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
joblib>=1.3.0
torch>=2.2.0
xgboost>=2.0.3
lightgbm>=4.3.0
shap>=0.45.0
sentence-transformers

# --- Visualization ---
matplotlib>=3.8.4
plotly>=5.21.0

# --- Frontend (Streamlit) ---
streamlit>=1.35.0
streamlit-lottie>=0.0.5
streamlit-option-menu>=0.3.13
extra-streamlit-components>=0.1.71

# --- AI & LLM (Full) ---
google-generativeai>=0.5.2
langchain-huggingface>=0.0.3
langchain-community>=0.0.38
langchain-core
langgraph>=0.0.10

# --- Utilities ---
fpdf2
pillow
feedparser>=6.0.10

# --- Data Pipeline (Orchestration) ---
apache-airflow>=2.8.0
pyspark>=3.5.0
